Updation of airline

+++ src/app/airline/airline.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { HttpService, Airline } from '../service/HttpService.service';
-import { Router} from '@angular/router';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-airline',
@@ -10,27 +10,30 @@ import { Router} from '@angular/router';
 export class AirlineComponent implements OnInit {
 
   airlines: any;
-  status:boolean;
-  tableStatus:boolean;
-  message:String;
-  airline: Airline = new Airline("","","");
+  status: boolean;
+  tableStatus: boolean;
+  updateStatus: boolean;
+  message: String;
+  airline: Airline = new Airline("", "", "");
 
-  constructor(private http : HttpService , private router: Router) { }
+  constructor(private http: HttpService, private router: Router) { }
 
   ngOnInit(): void {
     this.tableStatus = true;
     this.status = false;
+    this.updateStatus = false;
     this.message = "";
     this.getAirlineDetails();
   }
 
   getAirlineDetails() {
     this.tableStatus = true;
+    this.updateStatus = false;
     this.http.getAllAirlineDetails().subscribe(
       data => {
         console.log(data);
         this.airlines = data;
-        if(this.airlines.length == 0) {
+        if (this.airlines.length == 0) {
           this.tableStatus = false;
           this.message = "No Records Found!";
         }
@@ -38,20 +41,28 @@ export class AirlineComponent implements OnInit {
     )
   }
 
+  getSelectedAirline(airline: any) {
+    this.updateStatus = true;
+    this.airline = airline;
+  }
+
   deleteAirline(airline: Airline) {
-    this.http.deleteAirlineDetails(airline.name).subscribe(
-      data => {
-        console.log(data);
-        if(data == "Success") {
-          this.getAirlineDetails();
-        }
-      }, error => console.log(error)
-    )
+    if (confirm("Are you sure you want to cancel airline operation of " + airline.name + " ?")) {
+
+      this.http.deleteAirlineDetails(airline.name).subscribe(
+        data => {
+          console.log(data);
+          if (data == "Success") {
+            this.getAirlineDetails();
+          }
+        }, error => console.log(error)
+      )
+    }
   }
 
   registerAirline() {
     var flag = true;
-    if(this.airline.name == "") {
+    if (this.airline.name == "") {
       flag = false;
       alert("Please fill the Airline Name!");
     } else if (this.airline.address == "") {
@@ -61,20 +72,31 @@ export class AirlineComponent implements OnInit {
       flag = false;
       alert("Please provide 10 digit contact number!")
     }
-
-    if(flag) {
+    if (flag) {
       console.log(this.airline);
       this.airline.name = this.airline.name.toUpperCase();
-      this.http.registerAirlineDetails(this.airline).subscribe(
-        data => {
-          console.log(data);
-          this.airline = new Airline("","","");
-          this.getAirlineDetails();
-        }, error => {
-          console.log(error);
-          alert("The Airline Name is registered by other other. Please provide unique airline name!");
+      this.http.getAirlineByName(this.airline.name).subscribe(data => {
+        if (data != null) {
+          if (confirm("Are you sure you want to update the airline details of " + this.airline.name + " ?")) {
+            this.http.updateAirline(this.airline, this.airline.name).subscribe(data => {
+              console.log(data);
+              this.getAirlineDetails();
+              this.airline = new Airline("","","");
+            }, error => console.log(error));
+          }
+        } else {
+          this.http.registerAirlineDetails(this.airline).subscribe(
+            data => {
+              console.log(data);
+              this.airline = new Airline("", "", "");
+              this.getAirlineDetails();
+            }, error => {
+              console.log(error);
+              alert("The Airline Name is registered by other. Please provide unique airline name!");
+            }
+          )
         }
-      )
+      })
     }
   }
